backend/requirements.txt
fastapi==0.110.1
uvicorn==0.25.0
httpx==0.23.3
python-dotenv>=1.0.1
pydantic>=2.6.4
email-validator>=2.2.0
pyjwt>=2.10.1
passlib>=1.7.4
python-multipart>=0.0.9
typer>=0.9.0
WeasyPrint

# LLM API clients
anthropic>=0.22.0  # Pour l'API Claude
google-generativeai>=0.4.0  # Pour l'API Gemini
